+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # 0.0.166 (Feb 26, 2026)
 * Added `modules list` command to list modules in a Nullstone organization.
 * Fixed `nullstone run` usage text to properly show command args syntax (`[-- command [args...]]`).
+* Added `apps capabilities list` command to list capabilities in the app's template or app workspace.
+* Added `apps capabilities create` command to create capability in the app's template or app workspace.
+* Added `apps capabilities remove` command to remove capability in the app's template or app workspace.
 
 # 0.0.165 (Feb 18, 2026)
 * Fixed `modules generate` from failing during prompt.